feat(windows): share egui console UI and add native renderer, settings dialog, and menu
