+++ apps/desktop/Cargo.toml
@@ -38,6 +38,16 @@ libc = "0.2"
 cocoa = "0.26"
 objc = "0.2"
 
+[target.'cfg(target_os = "windows")'.dependencies]
+windows-sys = { version = "0.59", features = [
+  "Win32_Foundation",
+  "Win32_UI_WindowsAndMessaging",
+  "Win32_UI_Controls",
+  "Win32_UI_Input_KeyboardAndMouse",
+  "Win32_Graphics_Gdi",
+  "Win32_System_LibraryLoader",
+] }
+
 [package.metadata.bundle]
 name = "Sujay"
 identifier = "com.tnayuki.sujay"
